Add bookmarked articles
